Re-organize - break out production code and Treatment classes
Still a WIP
- need to build the things that were in guid_maps
- get tests passing
